Design fait, boutton _anuler modif_ ajouté, reste plus qu'à faire une modif à la fois

+++ src/components/DeleteArticle.js
@@ -1,21 +1,28 @@
-import React from 'react';
-import axios from 'axios';
+import React from "react";
+import axios from "axios";
 
 const DeleteArticle = (props) => {
-    const handleDelete = ()=>{
-      axios.delete('http://localhost:3003/articles/' + props.id)
-      .then((Response)=>{
-        props.func();
-      })
-    }
-    return (
-        <button onClick={()=>{
-            if(window.confirm('Voulez vous supprimer cette pub ?'))
-                handleDelete();
-        }}>
-            Delete Art
-        </button>
-    );
+  const handleDelete = () => {
+    axios
+      .delete("http://localhost:3003/articles/" + props.id)
+      .then((Response) => {
+        if (Response.status === 200) {
+          props.func();
+          window.alert("article Well deleted !");
+        } else {
+          window.alert("Bad deletion !");
+        }
+      });
+  };
+  return (
+    <button
+      onClick={() => {
+        window.confirm("Voulez vous supprimer cette pub ?") && handleDelete();
+      }}
+    >
+      Delete Article
+    </button>
+  );
 };
 
 export default DeleteArticle;
